+++ e2e/test_012_download_stable_zip_sig_screen.py
@@ -72,25 +72,6 @@ class TestDownloadStableZipSigScreen(GraphicUnitTest):
     @patch(
         "src.app.screens.base_screen.BaseScreen.get_locale", return_value="en_US.UTF-8"
     )
-    @patch("src.app.screens.base_screen.BaseScreen.redirect_error")
-    def test_fail_update_key(self, mock_redirect_error, mock_get_locale):
-        screen = DownloadStableZipSigScreen()
-        self.render(screen)
-
-        # get your Window instance safely
-        EventLoop.ensure_window()
-
-        # do tests
-        screen.update(name=screen.name, key="mock")
-
-        # patch assertions
-        mock_redirect_error.assert_called_once_with('Invalid key: "mock"')
-        mock_get_locale.assert_any_call()
-
-    @patch.object(EventLoopBase, "ensure_window", lambda x: None)
-    @patch(
-        "src.app.screens.base_screen.BaseScreen.get_locale", return_value="en_US.UTF-8"
-    )
     def test_update_locale(self, mock_get_locale):
         screen = DownloadStableZipSigScreen()
         self.render(screen)
