+++ src/app/screens/download_selfcustody_pem_screen.py
@@ -66,11 +66,8 @@ class DownloadSelfcustodyPemScreen(BaseDownloadScreen):
             )
             Clock.schedule_once(fn, 0)
 
-        self.debug(f"Bind {self.__class__}.on_trigger={on_trigger}")
-        setattr(self.__class__, "on_trigger", on_trigger)
-
-        self.debug(f"Bind {self.__class__}.on_progress={on_progress}")
-        setattr(self.__class__, "on_progress", on_progress)
+        setattr(DownloadSelfcustodyPemScreen, "on_trigger", on_trigger)
+        setattr(DownloadSelfcustodyPemScreen, "on_progress", on_progress)
 
         fn = partial(self.update, name=self.name, key="canvas")
         Clock.schedule_once(fn, 0)
